+++ core/src/lib.rs
@@ -3,6 +3,7 @@
 // TODO: #![warn(clippy::missing_panics_doc)]
 #![warn(clippy::clone_on_ref_ptr)]
 #![deny(trivial_numeric_casts)]
+#![forbid(unsafe_code)]
 
 use std::time;